Fix 7: ERC20 compliance
Although a minor problem, the decimals state variable should be defined with type uint8 to be compliant with ERC20.
Additionally, there is a restriction to the usage of approve intended to mitigate a problematic race condition, but it is not ERC20 compliant. Consider removing the line, and including an alternative mitigation.