Fix 9: Unbounded loops and Profit distribution
Unbounded loops
There are several loops in the contract which can eventually grow so large as to make future operations of the contract cost too much gas to fit in a block. Additionally, using unbounded loops incurs in a lot of avoidable gas costs for all token transactions.
As an example, there is a state variable allFNUHolders which holds an array with the addresses of all token holders. When an address becomes a token holder, it is added to this array. Whenever an address is left with balance zeroafter a transfer, the array is iterated over to find the address and remove it (removeAddress). If the array is large enough, the iteration could cost too much gas and it would be impossible to make a transfer function call.
Additionally to addAddress and removeAddress, the allFNUHolders array is iterated in the shareProfits function, to assign each token holder their corresponding profit in the allProfitHolders array. The same unbounded loop problem applies. This kind of imperative code-style is not a good fit for Ethereum smart contracts. 
To fix this problem the two arrays have to be removed, and the profit sharing feature completely redesigned.
One suggestion for implementing these features securely is calculating the profits “on demand” for each token holder, instead of ahead of time for all of them. This would require the token to support taking a snapshot of all balances at the time profits are loaded into the smart contract. It might be worth considering using the MiniMe Token for this.

Profit distribution can be reset by owner at any time
There is a function PrepareForProfitShare callable by the owner at any time, which resets the distribution of profits: it brings every token holder’s claimable profits down to zero. It’s not clear what purpose it serves. Consider removing it, as it harms the trustlessness of the contract.